factored out sticman into its own component
