+++ src/components/StickMan/styles.js
