--- src/screens/MainScreen/components/AddCartForm/components/ImagePicker/components/StickManPreview/index.js
@@ -1,14 +1,0 @@
-import React from "react";
-import { Container, HeadImg, BodyImg } from "./styles";
-import { StickManBody } from "../../../../../../../../constants";
-
-const StickManPreview = ({ headSrc }) => {
-  return (
-    <Container>
-      <HeadImg src={headSrc} alt="Preview Head" />
-      <BodyImg src={StickManBody} alt="Preview Body" />
-    </Container>
-  );
-};
-
-export default StickManPreview;
