+++ src/screens/MainScreen/components/Train/components/Cart/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Container, Name, CartContainer } from "./styles";
-import StickMan from "./components/StickMan";
+import StickMan from "../../../../../../components/StickMan";
 
 const Cart = ({ name, color, cartSize, facepic }) => (
   <Container cartSize={cartSize}>
